remove file Form.tsx, rename interface form to user, context
