+++ src/app/App.tsx
@@ -1,17 +1,17 @@
 import React, { useEffect, useState } from "react";
 import "../stylesheets/styles.css";
-import { BrowserRouter, Routes, Route, useLocation } from "react-router-dom";
-import { MyGlobalContext } from "./shared/contexts/cart.context";
+import { Routes, Route, useLocation } from "react-router-dom";
+import { MyGlobalContext } from "./shared/contexts/global.context";
 import { ICart } from "./shared/interfaces/cart";
 import { Home, Cart, Register } from "./pages/index";
 import { Header, Footer } from './shared/components/layout/index';
 import { getData, setData } from './shared/common/common';
-import { IForm } from './shared/interfaces/form';
+import { IUser } from './shared/interfaces/user';
 
 function App() {
   const location = useLocation();
   const [cart, setCart] = useState<ICart[]>(getData('cart', []));
-  const [user, setUser] = useState<IForm>(getData('user', {}));
+  const [user, setUser] = useState<IUser>(getData('user', {}));
   
   useEffect(() => {
     if (!getData('cart', null)) {
@@ -40,7 +40,7 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/cart" element={<Cart />} />
-        <Route path='/register' element={<Register />}/>
+        <Route path="/register" element={<Register />}/>
       </Routes>
       <Footer />
     </MyGlobalContext.Provider>
