+++ src/app/pages/Cart/partials/CartItem.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useGlobalContext } from '../../../shared/contexts/cart.context';
+import { useGlobalContext } from '../../../shared/contexts/global.context';
 import { ICart } from '../../../shared/interfaces/cart';
 import { Button } from '../../../shared/components/partials/index';
 import { formatFixed } from '../../../shared/common/common';
