+++ src/app/shared/contexts/global.context.tsx
@@ -1,26 +1,26 @@
 import React, { createContext, useContext } from 'react';
-import { ICart } from './../interfaces/cart';
+import { ICart } from '../interfaces/cart';
 import { getData } from '../common/common';
-import { IForm } from '../interfaces/form';
+import { IUser } from '../interfaces/user';
 
 const initialStateUser = {
-  email: "",
-  password: "",
-  confirmPassword: "",
+  email: '',
+  password: '',
+  confirmPassword: '',
 };
 
 interface GlobalContent {
   cart: ICart[];
   setCart: (cart: ICart[]) => void;
-  user: IForm;
-  setUser: (user: IForm) => void;
+  user: IUser;
+  setUser: (user: IUser) => void;
 }
 
 export const MyGlobalContext = createContext<GlobalContent>({
-  cart: getData("cart", []),
+  cart: getData('cart', []),
   setCart: (data: ICart[]) => data,
   user: initialStateUser,
-  setUser: (user: IForm) => user,
+  setUser: (user: IUser) => user,
 });
 
 export const useGlobalContext = () => useContext(MyGlobalContext);
