+++ src/app/shared/interfaces/user.ts
@@ -1,4 +1,4 @@
-export interface IForm {
+export interface IUser {
   email: string,
   password: string,
   confirmPassword: string
